requirements.txt
minizinc==0.8.0
Flask==3.0.0
python-dotenv==1.0.0